Take Recorder: Various fixes for recording livelink data.
Added missing GetMovieSceneSection() to LiveLinkTrackRecorder so that SetSectionStartTimecode is called properly.
RecordFrame should ExpandToFrame the exact frame, not the lower bound + frame.
Fix drawing collapsed keys losing precision.

#jira UE-87738
#rb none

[CL 11189751 by Max Chen in Dev-Tools-Staging branch]

+++ Engine/Source/Editor/Sequencer/Private/SSequencerSection.cpp
@@ -1411,7 +1411,7 @@ void SSequencerSection::PaintKeys( FSequencerSectionPainter& InPainter, const FW
 			// Generate the hull of frame numbers that contribute to this key so we can draw it enabled/disabled depending on whether it is outside of the valid range or not
 			TRange<FFrameNumber> KeyRange = TRange<FFrameNumber>::Empty();
 
-			float AverageKeyTime = 0.f;
+			double AverageKeyTime = 0.f;
 			int32 NumKeyTimes = 0;
 			int32 NumOverlaps = 0;
